# backend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Sports Management API
  description: |
    Comprehensive API for managing sports leagues, referee assignments, games, teams, and budgets.
    
    ## Authentication
    All endpoints (except `/health` and `/auth/login`) require Bearer authentication.
    Include the JWT token in the Authorization header: `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Login endpoint: 5 requests per minute
    - Registration endpoint: 3 requests per hour
    - Password reset: 3 requests per hour
    - Assignment operations: 100 requests per minute
    - General endpoints: 1000 requests per hour
    
    ## Error Handling
    All error responses follow a consistent format:
    ```json
    {
      "error": "Error message",
      "message": "Detailed description",
      "code": "ERROR_CODE"
    }
    ```
    
    ## Data Models
    The API follows a hierarchical structure:
    - Organizations contain Leagues
    - Leagues contain Teams
    - Games reference Teams and are assigned Referees
    - Budgets track financial aspects
    
  version: 1.0.0
  contact:
    name: Sports Management API Support
    email: support@sportsmanagement.app
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.sportsmanagement.app
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Team name
        league_id:
          type: string
          format: uuid
          description: Reference to league
        rank:
          type: integer
          minimum: 1
          description: Team ranking
        location:
          type: string
          description: Team's home location
        contact_email:
          type: string
          format: email
          description: Team contact email
        contact_phone:
          type: string
          description: Team contact phone
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - name
        - league_id
...
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error description
        code:
          type: string
          description: Error code for programmatic handling
...
security:
  - bearerAuth: []
paths:
...
  /teams:
    get:
      tags:
        - Teams
      summary: Get all teams
      description: Retrieve all teams with optional filtering
      parameters:
        - name: league_id
          in: query
          description: Filter by league ID
          schema:
            type: string
            format: uuid
        - name: organization
          in: query
          description: Filter by organization
          schema:
            type: string
        - name: age_group
          in: query
          description: Filter by age group
          schema:
            type: string
        - name: gender
          in: query
          description: Filter by gender
          schema:
            type: string
            enum: [Boys, Girls, Mixed]
        - name: season
          in: query
          description: Filter by season
          schema:
            type: string
        - name: search
          in: query
          description: Search by team name
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Teams retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

    post:
      tags:
        - Teams
      summary: Create new team
      description: Create a new team (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Team'
            example:
              name: "Lightning Bolts"
              league_id: "123e4567-e89b-12d3-a456-426614174000"
              rank: 1
              location: "North Field Complex"
              contact_email: "coach@lightning.com"
              contact_phone: "555-0123"
      responses:
        '201':
          description: Team created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  team:
                    $ref: '#/components/schemas/Team'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /teams/{id}:
    get:
      tags:
        - Teams
      summary: Get specific team
      description: Retrieve a specific team by ID
      parameters:
        - name: id
          in: path
          required: true
          description: Team ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Team retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  team:
                    $ref: '#/components/schemas/Team'
        '404':
          description: Team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Teams
      summary: Update team
      description: Update an existing team (admin only)
      parameters:
        - name: id
          in: path
          required: true
          description: Team ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Team'
      responses:
        '200':
          description: Team updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  team:
                    $ref: '#/components/schemas/Team'
        '404':
          description: Team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Teams
      summary: Delete team
      description: Delete a team (admin only)
      parameters:
        - name: id
          in: path
          required: true
          description: Team ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Team deleted successfully
        '404':
          description: Team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Games
...